requirements.txt
alembic>=1.18.4
asyncpg>=0.31.0
bcrypt==4.0.1
deepgram-sdk==5.2.0
fastapi>=0.131.0
groq>=1.0.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
pydantic-settings>=2.13.1
pymupdf>=1.27.1
python-docx>=1.2.0
python-dotenv>=1.2.1
python-jose>=3.5.0
python-multipart>=0.0.20
sounddevice>=0.5.5
soundfile>=0.13.1
sqlalchemy>=2.0.46
sqlalchemy-utils>=0.42.1
structlog>=25.5.0
uvicorn>=0.41.0
websockets>=16.0
livekit-api>=0.8.0